added key_value_unsafe.c ; implemented workout persistence

+++ src/lib/persistor.h
@@ -2,7 +2,7 @@
 
 void persist_read_long_string(uint32_t index, char *dest);
 
-int persist_read_long_string_length(uint32_t index);
+size_t persist_read_long_string_length(uint32_t index);
 
 bool persist_write_long_string(uint32_t index, char *data);
